+++ include/prime.h
@@ -1,4 +1,5 @@
 #include <stdio.h>
+#include <stdlib.h>
 #include <math.h>
 
-int check_prime(unsigned long long n);
+int check_prime(int n);
